Bug Fix
There was a bug in generate dummies caused by a trailing space in the data (in Gender column). This meant that this specific question would crash the app, as it couldn't find dummies ef the same gender (which was unique because of the typo). I corrected the data, and seized the opportunity to make the app slightly more robust to this kind of inconsistencies in the data.